Implement relative date component
I've been fooling around for about an hour with this and did not commit
in frustration of not getting it to do what I wanted it to do. In the
end decided to drop it for now.

Also not going to work in 'sessions' anymore. As I've noticed I would
not start on this when I don't feel like I've enough time on my hands to
do an actual session.